migrate to 32-bit CSRs
Contrary to popular belief, 32-bit CSRs actually seem to take
less resources than 8-bit CSRs. The design from the previous
commit compiles to 5911 slices; with 32-bit CSRs turned on,
it compiles to 5832 slices.

Overall, congestion is slightly lower for the 32-bit CSR design,
but neither design is highly congested so it's difficult to
draw conclusions -- the analysis tool will report no congestion
results on the default level for the final design, only on
the initial estimated design.

Perhaps one of the most telling analysis is the logic level
distribution. For the 32-bit CSR, the distribution looks like this:

levels  0   1   6   8   9   11   12   13   15
count   0   96  2   37  31  54   491  276  1
(988 total)

50% of paths are 12 and below, 99% are 13 and below.

For 8-bit CSR, the distribution looks like this:

levels  0   1    2   4   5   6   7   8   9    11   12   13   14   15
count   0   119  1   1   1   1   31  53  146  1    38   243  282  71
(988 total)

In this case, 50% of the paths are 13 and above, with 35% 14+ deep.

As betrusted-soc makes heavy use of 16 and 32-bit CSRs, going
with an 8-bit CSR bus means 2-4x the number of individual registers
to decode and address. This increases the logic depth by 1-2 extra
levels, in exchange for a reduction in routing resource usage.

Given the relative lack of congestion in the design, reducing
logic levels and slice usage by going to 32-bit CSRs seems to
be the right choice.

Furthermore, by optimizing the accessors for 32-bit CSRs to be
a single read, code execution runs faster, especially for
hardware routines that rely heavily upon CSR accesses.