add routines to test XADC option for audio input
XADC actually performs OK for audio input, except that
it lacks an anti-aliasing filter, and so you can hear some
aliasing artifacts. The noise level is worse than the
pure codec solution as well, but this is expected.

Overall, we will probably go with the pure codec option;
it will be a better user experience, cheaper, and the
attack surface of the codec inside the trusted domain is
likely quite manageable.