ram standby testing
Write a pattern into 8MiB of RAM, and check it. This was
used to confirm that during power-off, the RAM retains
data as intended. The power-off retention of data is intended
to reduce boot time of the FPGA, so that we can be very
aggressive about powering the FPGA off to extend battery
life.

As noted elsewhere, the retention of RAM does mean we have
to lean strongly toward implementing encrypted RAM to
defeat readout/upset attacks of the system. But one thing
at a time.