fix TDO sampling timing problem
In the code I had originally sampled TDO before doing a TCK,
which is the correct idea, except I forgot that the hardware
enforces this as well (by sampling TDO before allowing
TCK to change). So in fact, the correct program order is
to read TDO after triggering the TCK.

The reason this is done in hardware is it allows the hardware
to enforce setup/hold timing around TDO sampling, instead of
relying on a software loop.