add a signal-rs package to the firmware
This is adapted from chouchand/signal-rs.

This is not meant to reflect an accurate implementation of
Signal. In particular, I had to rewrite the HKDF api, and
I am using SHA512 but it looks like the Signal "official" package
uses SHA256 (it would be pretty easy to backfill this implementation with
SHA256).

However, the intent of this package is to do the double-ratchet
computation with the crypto primitives we intend to deploy, so
that we can end-to-end test the correct hardware integration of all the
hardware crypto primitives. There is a double-ratchet test routine
that we grab out of signal-rs to do this.