add a rule to make all multiregs false paths
Multiregs are, by definition, false paths -- they go between
clock domains and timing is intended to fail between them, but
the timer can get choked on these if they are not explicitly declared.

Add two rules that take the '0' register in a MultiReg chain and
make all signals coming from it (sys->other domain) and signals
going to it (other->sys) false paths. This removes unecessary
constraints and allows the router to do a better, faster job.