stretch out the console wakeup pulse
UART characters take a long time to come in, and characters that
are sparse and have just an MSB set will fail to keep the CPU
awake long enough to register the incoming character. this adds
a one-shot that keeps the CPU awake for about 2x the length of
a UART character.