tweak optimizer settings to shrink down the bootloader
This turns out to be the thing that gets the code to fit again.
In particular, `opt-level = 'z'` was generating much larger code
than `opt-level = 's'`. The rust docs say you have to play with
these a bit, because in some cases an opt-level of 3 can produce
smaller code than opt-level s. Anyways it seems the codegen
backend shifted somewhere recently and for rustc 1.60 this is
the setting we need to get this to fit.