fix the location of the vexriscv_debug
LiteX (or something in between) resets our preference for where the
CPU debug block should go to 0xf00f0000, which doesn't work for us
because we have 4k pages for our CSRs and we encode CSRs into that
space. I think no other builds encounter this problem because nobody
else does that.

Anyways, recode the ACL for the USB using the hard-coded parameters.
this parameter is also hard-coded in several spots through the design
file, fwiw.