add force_reset option
the auto-reset is now to spec (100ms) but it's short enough that
it looks like some Linux implementations view this as a glitchy
USB port and eventually the kernel decides to just ignore the port
altogether.

Add a force_reset option so we can hold the port in reset while the
switchover happens, and hold it for a couple seconds so that the
kernel doesn't conclude we have a bad cable or something like
that.