pull Vex configuration into the root project
The pythondata-cpu-vexriscv build infrastucture is too inflexible
and out of dated for the configuration that we're using.

Pulling the Vex config into the local build allows us to target
the dev tree of SpinalHDL and pull in more patches, instead of
having to wait for a release cycle on the Spinal side.

this is tested and seems to be baseline with previous functionality.

This also has the benefit that the material source files are
now more intimately linked to this repo, instead of splattered
into submodules everywhere.