add fifo joining, fix initial values on control registers
FIFO joining implementation probably has an extra cycle of latency
over what's in the rp2040, because we use a mux to chain the two fifos
without re-doing the counters to have a full view of all the entries.
Thus, the joined FIFO effectively adds a cycle of latency versus
what is drawn in the rp2040 manual diagram.

I think this is probably fine in practice?

also fixed IV settings for control/config registers