align immediate instructions to the action register pulse edge
previously a mis-alignment meant that the immediate instruction
that is pending (and possibly stalling) could be swapped
out while a new immediate instruction is issued.

in other words, the timing of the "do the immediate instruction"
and the immediate instruction itself was imprecise; the action
pulse arrived a bit late.

This synchronizes the two together, at an expense of an extra
cycle latency for immediate instruction execution and a bank
of registers.