make sm private again, add rx_level routine
sm is now a private variable again, and it is accessed by
a function that resolves it to a bitmask. This helps to make
code a bit clearer.

A todo note was added to split out program loading to its own
ball of mutable state separate from the SM's, because program
loading is actually a cross-SM activity.

rx_level read routine added.