add asserts to cause the test bench to terminate on failure
there are so many vectors now in the simulation that it's easy
to miss a failure buried in the noise.

Note that this is not iron-clad, some tests can be easily spoofed.
In particular, the stick_bit() test requires a manual review of
the outputs (it also tests very rapid state machine updates, so
we can't use PIO readback to confirm the results, we'd miss
cycles).

Also, the SPI test doesn't catch errors in output wiring
because the loopback pin is any pin that is wired to the
output.