work around Litex bug where AXI-lite registers don't have a one-wide WE pulse
the WE pulse (confusingly named, it's the read strobe) is two-wide,
violating the CSR spec, due to some implementation details of
the axi-lite to CSR converter.

this adds a rising edge detector so as to work around this.

It's not an issue for the SoC tape-out, since this is specific
to the LiteX harnessing around the SoC code itself.