strip out dma bypass on push/pull; add margins instead
Instead of trying to close the DMA request loop in a single
cycle (which seems impossible anyways, given the multiplicity
of clocking options available), we add a "margin" register that
allows us to specify from 0-3 items ofr "margin" on the IRQ
signal that would be routed to a DMA engine to request filling
or emptying of FIFOs.

Typically, you would only set this to "1", and the reason is
that sometimes the DMA fill or drain can be imprecise in that
the IRQ that initiates this may remain asserted long enough
after the bus operation happens that the DMA engine would
initiate an additional cycle (this is because the actual
bus write itself has to go through a 2-stage synchronizer,
which can be long enough a delay that the FIFO won't see
the write until the DMA is already preparing the next transfer).

Thus for a streaming output or input, the "margin" allows
for the DMA engine to sometimes write an extra element or
read an extra element without losing data.

Note that this margin is *only* applied to the IRQ line,
the actual full/empty signals to the status outputs and the
level signals are not affected.

If precise termination of a data stream is desired, then,
on the last transfer, one may need to add some code to
query if (for example) the rx_level is not 0, and read
the straggler out.

Generally, if a TX has a fixed length stream, the DMA
engine would terminate correctly without an extra flush
mechanism required, since the DMA should terminate
with the refill IRQ still active (the actual termination
happened because the length was reached, not becaause
things were full).

This commit also includes some test cases that check
for the case of margin=1 the fifos have the correct levels
for both the single and joined cases.