fix imprecise CTRL register, and make irq_stb level based
- irq_stb must be level-based, because with clkdiv = 1.0,
successive _stb operations get dropped. Instead, we make
the _stb generated by imm || penable (exec is encompassed
in penable) so we don't have to be sensitive to edges

- clean up imprecise CTRL register. In the case that the CPU
is totally synchronous with PIO, you can assume that a write
to clkdiv_restart or restart happens cleanly before the next
write. However, in our case, the CPU can run much, much faster
than the PIO block. There's a few ways to try to handle this,
and I opt for the simplest way: the semantics of the register
are clear-on-write, but the CPU has to read back the register
to ensure the clear has happened before moving on to the next
operation.

The alternative would be to insert some sort of FIFO/coalescing
structure that would try to capture outstanding concurrent writes
to the block and ensure they are applied in sequence, but eventually
you could exhaust the depth of the FIFO and so forth. Generally
I think the CTRL register ops are hidden behind API calls
that can implement the readback, so pushing this up into the
CPU stack to handle instead of adding more hardware and possible
sources of bugs is my preferred approach.