fix empty conditions on joined fifos
the fifo is empty whenever the outermost fifo providing the
pull data is empty.

anding the two signals together causes a problem where data
propagataing down the inner fifo makes the empty look not
empty, which can trigger an immediate pull of the fifo,
but this yields bogus data because it hasn't reached the
fifo repsonding to the pull!