update common simulation bench to target both verilator and xsim
Requires a wrapper script. The wrapper script is responsible
for building the SVD file and firmware image, and then starting
the simulator (requires two sucessive calls to the cram_soc.py script).

`verilate.sh` runs verilator, works in linux
`xsimulate.ps1` runs xsim, runs on windows

next up is looping back to the arty configuration and having
it rely on the "common" configuration to build the synthesizeable
FPGA bitstream.