commit changes to config to target the soc build
Now, the distinction between SoC vs FPGA is as follows

- SoC builds closely match the production hardware, including
proprietary blocks that we don't have in our tree. This means
it is increasingly difficult to simulate SoC builds locally.
However, it means that simulations on the production cluster
are more accurate, so this is a reasonable trade-off when pushing
toward tape-out.

- FPGA builds track what can go onto the Arty platform. This
tries to be somewhat close to the SoC builds but things like
the clock generator and some of the PHY blocks are just going
to be different because these are platform-specific IP blocks.