add external mailbox and modify interrupt triggers
interrupt triggers now support selectable edge/level triggers
on a per-interrupt basis

can also now select an external mailbox as the source for
the mailbox protocol.

The loopback path is kept because in the case that we want to
use the mbox for interprocess entirely on the RV32, it's faster
with the local mbox. The external mbox is useful because the
M7 can access it, allowing potential inter-CPU coordination
(however this is very dangerous because there is no hardware guarantee
that the M7 can't tamper with the RV32's memory space, one would
have to configure the MPUs to lock things down and run out of
TCM, but, this might be helpful anyways for offloading untrusted DSP
operations to the much faster CPU).