add gpio snap-to-quantum register option
this allows users to specify if the gpios should update immediately,
or if they should only update whenever a quantum increment happens;
and if so, which quantum to use.

Only one quantum can be selected across the entire bank with the
current configuration. I suppose maybe this reduces the possibility
of having multiple engines doing different I/O operations but this
would only be in the edge case that you had a compound I/O function
that synchronized across multiple cores handling different pins,
and then you'd have two of those loaded into the same BIO.

In other words, in the case that you had four cores each handling
its own I/O function with its own set of pins, and they are
unrelated to each other, the quantum function within each core
is sufficient.

The only case where you'd want to snap to a "global" quantum is
if you had more than one core managing a common set of GPIO pins,
and for some reason the helper core wasn't tightly synced to
the primary core.

Basically, it's a convenience to have the snap, but, it's also
something that's kind of edge-casey, so let's not burn the hardware
on giving every pin an independently configurable quantum snap.