change behavior of pindir toggle & normalize gpio mapping
- the gpio mapping was reversed because there's a bug in the SoC
integration but we shouldn't bake that into this code. let's
fix it in the SoC.

- the pindir clear behavior clears on `1` as opposed to `0` like
the pinout behavior.

The pinout has a somewhat counter-intuitive behavior because
we want to be able to do a fast toggle for a SPI interface;
if we had to invert the bits before sending to pins it'd add
an instruction to the pipe and hurt max speed.