fix regressions incurred during RAM refactor
BIO-BDMA code space for each CPU is now in a separate private
RAM. Refactor the code around the specific RAM macro that we
intend to use for production, and also refactor the bus interface
to have a page-oriented APB interface for each page.

The page-oriented interface is a bit annoying to hook up but the
APB space is pre-decoded into a set of pages by a mux in the system
already, so this allows for direct compatibility with that harness.