add bank aliases; add external dmareq interfaces
also changed event response behavior to trigger if any bit
of the bit mask is set; this is more flexible as it allows
a core to respond to multiple events. if event-response behavior
is needed that is gated on both events being set, this can be
handled in software with little loss of timing accuracy.

bank aliases allow us to map registers to different pages for processes
to have exclusive access to them. The registers are divided across
fifo0-3; every process can view the fifo levels and event status bits
(and can also set arbitrary event bits), but only one process can
actually stick data into or read data out of a fifo.

If a developer tries to have both the mapped alias and the main
control thread access a fifo simultaneously, the behavior is undefined.

just don't do that, you don't need to. whatever you thought would
be clever with that is a bad idea.

also add a method for taking all of the IRQs in the system and
condensing them down into event-set bits. This allows the BIO-BDMA
to be a repsonder to an IRQ from other peripherals without having
to bother the main CPU.