add optional DMA filters to BIO-BDMA
up to four base/bound regions can be permitted at one time
when filtering is enabled. The intention is that generally one
would block read access from ROM (where secrets may be kept),
the configuration of the BIO peripheral itself (so the CPU can't
use its own code to turn off filtering), and then an spare range
which can be used to help confine a RAM target for DMA cycles.