add timing options & multiplier configuration
A set of `defines are now available at the top of bio-bdma.sv block
that allows us to tune timing closure vs IPC of the block, once
we have the back-end synthesis and timing estimators in place.

Multiplier option is also integrated; it doesn't seem to add too
much to the area or cycle time, and the benefit could be pretty
interesting in terms of delegating a MAC-style dot product while
doing a DMA to any of the BIO cores.